refactor: extract icon picker to shared component, Puck re-exports from shared
