+++ src/components/shared/IconPicker/index.ts
@@ -1,0 +1,4 @@
+export { IconPicker } from './IconPicker';
+export { IconRenderer, iconToHtml } from './IconRenderer';
+export { searchIcons, getLucideIcons, getHeroicons, getEmojis } from './icon-catalog';
+export type { IconEntry } from './icon-catalog';
